refactor(frontend): extract getInitialInputValue from AreaM2EditCell
Move the pure getInitialInputValue helper out of the AreaM2EditCell
component into a dedicated areaM2EditCellValue module and add unit tests
covering raw-string passthrough, numeric formatting, fallback handling,
and rounding. No behavior change.
